Check whether retrieved IP is in a Private Range
If the IP we get back from the interface is within a Private Range, it should not probably be used to update a public IP record as it would be unresolvable externally. Check the IP we retrieved against these known blocks and if it falls within one, use an external service (in this case https://ifconfig.me/ip) to retrieve the true public IP. A future enhancement could read a preferred service from an environment variable. Expected return format is simply a string containing the public IP address.